pyproject.toml
# Define the build system configuration.
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Define the project configuration.
[project]
name = "tung_playground"
description = """
Make your AI generated heroes live in the simulator.
"""
dynamic = ["version"]
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.10"
dependencies = [
    "mujoco",
    "numpy",
    "opencv-python",
    "pillow",
    "pyyaml",
    "trimesh",
    "open3d",
    "scipy",
    "pydantic>=2.0",
    "rich",
    "typer",
    "requests",
    "aiohttp",
    "asyncio-mqtt",
    "pathlib",
    "dataclasses-json"
]
license-files = ["LICENSE"]
authors = [
    {name = "Chenyu Cao", email = "ruziniuuuuu@gmail.com"}
]
maintainers = [
    {name = "Chenyu Cao", email = "ruziniuuuuu@gmail.com"}
]
keywords = [
    "mujoco",
    "robotics",
    "3d-generation",
    "aigc",
    "simulation",
    "reinforcement-learning"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "isort",
    "mypy",
    "pre-commit"
]
ml = [
    "torch",
    "torchvision",
    "transformers",
    "diffusers",
    "accelerate"
]
vis = [
    "viser",
    "yourdfpy"
]
isaac = [
    # "omni-isaac-lab"  # Commented out - not available in package registry
]

[tool.hatch.version]
path = "src/tung_playground/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/tung_playground"]

[tool.hatch.metadata]
allow-direct-references = true